Rename task package
