+++ src/main/java/seedu/address/logic/parser/CreateStudentCommandParser.java
@@ -19,7 +19,7 @@ import seedu.address.model.profbook.Email;
 import seedu.address.model.profbook.Name;
 import seedu.address.model.profbook.Phone;
 import seedu.address.model.profbook.Student;
-import seedu.address.model.taskmanager.TaskList;
+import seedu.address.model.task.TaskList;
 
 /**
  * Parses input arguments and creates a new CreateStudentCommand object
