+++ src/main/java/seedu/address/model/util/SampleProfBook.java
@@ -12,8 +12,8 @@ import seedu.address.model.profbook.Name;
 import seedu.address.model.profbook.Phone;
 import seedu.address.model.profbook.Root;
 import seedu.address.model.profbook.Student;
-import seedu.address.model.taskmanager.Task;
-import seedu.address.model.taskmanager.ToDo;
+import seedu.address.model.task.Task;
+import seedu.address.model.task.ToDo;
 
 /**
  * Sample ProfBook data.
