+++ src/main/java/seedu/address/storage/JsonAdaptedGroup.java
@@ -18,10 +18,10 @@ import seedu.address.model.id.StudentId;
 import seedu.address.model.profbook.Group;
 import seedu.address.model.profbook.Name;
 import seedu.address.model.profbook.Student;
-import seedu.address.model.taskmanager.Deadline;
-import seedu.address.model.taskmanager.Task;
-import seedu.address.model.taskmanager.TaskList;
-import seedu.address.model.taskmanager.ToDo;
+import seedu.address.model.task.Deadline;
+import seedu.address.model.task.Task;
+import seedu.address.model.task.TaskList;
+import seedu.address.model.task.ToDo;
 
 /**
  * A class to adapt a Group object into a format suitable for JSON storage.
