+++ src/test/java/seedu/address/logic/commands/CreateGroupCommandTest.java
@@ -23,7 +23,7 @@ import seedu.address.model.profbook.Group;
 import seedu.address.model.profbook.Name;
 import seedu.address.model.profbook.Root;
 import seedu.address.model.profbook.Student;
-import seedu.address.model.taskmanager.TaskList;
+import seedu.address.model.task.TaskList;
 
 public class CreateGroupCommandTest {
