+++ src/test/java/seedu/address/logic/commands/CreateTodoCommandTest.java
@@ -30,9 +30,9 @@ import seedu.address.model.profbook.Group;
 import seedu.address.model.profbook.Name;
 import seedu.address.model.profbook.Root;
 import seedu.address.model.profbook.Student;
-import seedu.address.model.taskmanager.Task;
-import seedu.address.model.taskmanager.TaskList;
-import seedu.address.model.taskmanager.ToDo;
+import seedu.address.model.task.Task;
+import seedu.address.model.task.TaskList;
+import seedu.address.model.task.ToDo;
 import seedu.address.testutil.StudentBuilder;
 
 public class CreateTodoCommandTest {
