+++ src/test/java/seedu/address/logic/commands/MoveStudentToGroupCommandTest.java
@@ -21,7 +21,7 @@ import seedu.address.model.profbook.Group;
 import seedu.address.model.profbook.Name;
 import seedu.address.model.profbook.Root;
 import seedu.address.model.profbook.Student;
-import seedu.address.model.taskmanager.TaskList;
+import seedu.address.model.task.TaskList;
 
 public class MoveStudentToGroupCommandTest {
     public static final String ERROR_MESSAGE_INCORRECT_DIRECTORY = "Directory is invalid";
