+++ src/test/java/seedu/address/statemanager/ChildOperationTest.java
@@ -27,7 +27,7 @@ import seedu.address.model.profbook.Name;
 import seedu.address.model.profbook.Root;
 import seedu.address.model.profbook.Student;
 import seedu.address.model.profbook.exceptions.DuplicateChildException;
-import seedu.address.model.taskmanager.TaskList;
+import seedu.address.model.task.TaskList;
 import seedu.address.testutil.StudentBuilder;
 
 public class ChildOperationTest {
